added ci tests

+++ tests/test_python_api.py
@@ -12,8 +12,10 @@ from cad_to_h5m import cad_to_h5m
 class TestExtrudeCircleShape(unittest.TestCase):
 
     def setUp(self):
-        url = 'https://github.com/Shimwell/fusion_example_for_openmc_using_paramak/archive/refs/tags/v0.0.1.tar.gz'
-        urllib.request.urlretrieve(url, 'v0.0.1.tar.gz')
+        
+        if not Path('dagmc_small_faceting_tolerance.h5m').is_file():
+            url = 'https://github.com/Shimwell/fusion_example_for_openmc_using_paramak/archive/refs/tags/v0.0.1.tar.gz'
+            urllib.request.urlretrieve(url, 'v0.0.1.tar.gz')
 
         tar = tarfile.open('v0.0.1.tar.gz', "r:gz")
         tar.extractall()
@@ -21,7 +23,8 @@ class TestExtrudeCircleShape(unittest.TestCase):
 
 
     def test_h5m_file_creation(self):
-        """Checks that a h5m file is created from stp files"""
+        """Checks that a h5m file is created from stp files when make_watertight
+        is set to false"""
 
         os.system('rm dagmc.h5m')
 
@@ -57,12 +60,6 @@ class TestExtrudeCircleShape(unittest.TestCase):
                 }
             ],
             h5m_filename='dagmc.h5m',
-            # cubit_path='/opt/Coreform-Cubit-2021.5/bin/',
-            # surface_reflectivity_name='reflective',
-            # merge_tolerance=1e-4,
-            # cubit_filename='dagmc.cub',
-            # geometry_details_filename='geometry_details.json',
-            # faceting_tolerance=1.0e-2,
             make_watertight=True,
         )
 
@@ -80,14 +77,41 @@ class TestExtrudeCircleShape(unittest.TestCase):
                     'material_tag': 'mat1'
                 }
             ],
-            # h5m_filename='dagmc.h5m',
-            # cubit_path='/opt/Coreform-Cubit-2021.5/bin/',
-            # surface_reflectivity_name='reflective',
-            # merge_tolerance=1e-4,
             cubit_filename='dagmc.cub',
-            # geometry_details_filename='geometry_details.json',
-            # faceting_tolerance=1.0e-2,
-            # make_watertight=True,
         )
 
         assert Path('dagmc.cub').is_file()
+
+
+    def test_faceting_tolerance_increases_file_size(self):
+        """Checks that a h5m file is created from stp files"""
+
+        os.system('rm *.h5m')
+
+        cad_to_h5m(
+            files_with_tags=[
+                {
+                    'filename':'fusion_example_for_openmc_using_paramak-0.0.1/stp_files/blanket.stp',
+                    'material_tag': 'mat1'
+                }
+            ],
+            h5m_filename='dagmc_default_faceting_tolerance.h5m',
+            faceting_tolerance=1.0e-2,
+        )
+        
+        assert Path('dagmc_default_faceting_tolerance.h5m').is_file()
+
+        cad_to_h5m(
+            files_with_tags=[
+                {
+                    'filename':'fusion_example_for_openmc_using_paramak-0.0.1/stp_files/blanket.stp',
+                    'material_tag': 'mat1'
+                }
+            ],
+            h5m_filename='dagmc_small_faceting_tolerance.h5m',
+            faceting_tolerance=0.5e-2,
+        )
+
+        assert Path('dagmc_small_faceting_tolerance.h5m').is_file()
+
+        assert Path('dagmc_small_faceting_tolerance.h5m').stat().st_size > Path('dagmc_default_faceting_tolerance.h5m').stat().st_size
